refactor registration routing
